Correccion

+++ Backend/src/main/java/pe/edu/upc/backend/repositories/IRespuestaRepository.java
@@ -3,5 +3,5 @@ package pe.edu.upc.backend.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pe.edu.upc.backend.entities.Respuesta;
 
-public interface IRespuestaRepository extends JpaRepository<Respuesta, Long> {
+public interface IRespuestaRepository extends JpaRepository<Respuesta, Integer> {
 }
