+++ Backend/src/main/java/pe/edu/upc/backend/repositories/IUsuarioRutaRepository.java
@@ -3,5 +3,5 @@ package pe.edu.upc.backend.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pe.edu.upc.backend.entities.UsuarioRuta;
 
-public interface IUsuarioRutaRepository extends JpaRepository<UsuarioRuta, Long> {
+public interface IUsuarioRutaRepository extends JpaRepository<UsuarioRuta, Integer> {
 }
